refactor(agent): separate native command transport
